+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- Maven plugin example: [Kotlin](Kotlin%20Plugins/002%20MavenPlugin)
 - Simple Hello World example: [Java](Java%20Plugins/001%20Simple%20Hello%20World) - [Kotlin](Kotlin%20Plugins/001%20Simple%20Hello%20World)
 
 [Unreleased]: https://github.com/PlateStack/PlateStack/compare/v0.0.0...HEAD